% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercises.R
\name{build_answers_text}
\alias{build_answers_text}
\title{Builds text for answers}
\usage{
build_answers_text(text1, text2, text3)
}
\arguments{
\item{text1}{tibble for alternative 01}

\item{text2}{tibblefor alternative 02}

\item{text3}{tibble for alternative 03}
}
\value{
a list
}
\description{
Used in eoc TRUE/FALSE exercises
}
\examples{
text1 <- dplyr::tibble(text = c('O R é uma plataforma de programação madura e estável;',
'O R foi desenvolvido em 2018 e é um projeto inovador e instável;'),
sol = c(TRUE, FALSE))

text2 <- dplyr::tibble(text = c('O RStudio é uma interface ao R, aumentando a produtividade do analista;',
'O RStudio é uma linguagem de programação alternativa ao R;'),
                      sol = c(TRUE, FALSE))

text3 <- dplyr::tibble(text = c('O R tem compatibilidade com diferentes linguagens de programação;',
'O R não tem compatibilidade com diferentes linguagens de programação;'),
                      sol = c(TRUE, FALSE))

l <- build_answers_text(text1,
text2,
text3)
}
