% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercises.R
\name{exercises_build}
\alias{exercises_build}
\title{Builds eoc exercises}
\usage{
exercises_build(folder_in, type_doc)
}
\arguments{
\item{type_doc}{type of document (from knitr::pandoc_to())}

\item{files_in}{files to render exercises}
}
\value{
nothing
}
\description{
Builds eoc exercises
}
\examples{
exercises_build(exercises_dir_get(exercises_dir_list()[1]),
 type_doc = "html")
}
