% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exercises_get_img}
\alias{exercises_get_img}
\title{Returns local dir of pkg}
\usage{
exercises_get_img(img_file)
}
\arguments{
\item{img_file}{name of img file}
}
\value{
a path
}
\description{
Returns local dir of pkg
}
\examples{
exercises_get_img("exercises-latinometrics-instagram_01-SOLUTION-20220927.png")
}
