% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exercises_print_img}
\alias{exercises_print_img}
\title{Prints imgs in exercises}
\usage{
exercises_print_img(img_file, this_caption)
}
\arguments{
\item{img_file}{name of img file}

\item{this_caption}{caption of img}
}
\value{
nothing
}
\description{
Prints imgs in exercises
}
\examples{
exercises_print_img("exercises-latinometrics-instagram_01-SOLUTION-20220927.png", "NONE")
}
