% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_fcts.R
\name{vdr_get_links_book}
\alias{vdr_get_links_book}
\title{Gets official links from book}
\usage{
vdr_get_links_book()
}
\value{
A list with links
}
\description{
Use this function in the book so that all links are updated in a single location.
}
\examples{

print(vdr_get_links_book())
}
