% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{versions_string}
\alias{versions_string}
\title{Formats versions for plots (used in cowplot::plot_grid())}
\usage{
versions_string(n_plots)
}
\arguments{
\item{n_plots}{number of plots}
}
\value{
a string
}
\description{
Formats versions for plots (used in cowplot::plot_grid())
}
\examples{
versions_string(10)
}
