% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_pkg_citation}
\alias{format_pkg_citation}
\title{Produces package citation string}
\usage{
format_pkg_citation(pkg)
}
\arguments{
\item{pkg}{a pkg availabe locally or github}
}
\value{
a string in rmarkdown
}
\description{
Formats package as `pkg` and, when called the first time, adds reference as
[@R-pkg]
}
\examples{
format_pkg_citation("dplyr")
}
