% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_fcts.R, R/links.R
\name{links_get}
\alias{links_get}
\title{Gets official links from book}
\usage{
links_get()

links_get()
}
\value{
A list with links

A list with links
}
\description{
Use this function in the book so that all links are updated in a single location.

Use this function in the book so that all links are updated in a single location.
}
\examples{

print(links_get())

print(links_get())
}
